Uniform project structure
Chapter 4.

+++ Part I - The Basics/Chapter 4 - Computation/Drill/src/ch_4_drill_4.cpp
@@ -1,15 +1,15 @@
 //Copyright 2016 Vinicius Sa (viniciusjavs@gmail.com)
-//Chapter 4, Drill 3
+//Chapter 4, Drill 4
 /*
- * Program that reads two ints continously into a while-loop
+ * Program that reads two doubles continously into a while-loop
  * and prints the smaller and larger values, or if they are equals.
  * Entering '|' exits the program.
 */
-#include "../../../include/std_lib_facilities.h"
+#include "std_lib_facilities.h"
 
 int main()
 {
-  int a, b;
+  double a, b;
   while (cin >> a >> b) {
     if (a > b) {
       cout << "the smaller value is: " << b << '\n';
